mnist-pytorch/requirements.txt
torch
torchvision
numpy
pandas
matplotlib
seaborn
scikit-learn
joblib
tqdm
Pillow
